# docs/swagger.yaml
basePath: /
definitions:
...
  gos_internal_domain_application.GitOpsBranchMapping:
    properties:
      branch:
        type: string
      env_code:
        type: string
    type: object
  gos_internal_domain_application.ReleaseBranchOption:
    properties:
      branch:
        type: string
      name:
        type: string
    type: object
  internal_interfaces_http.ApplicationDataResponse:
    properties:
      data:
        $ref: '#/definitions/internal_interfaces_http.ApplicationResponse'
    type: object
  internal_interfaces_http.ApplicationListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/internal_interfaces_http.ApplicationResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  internal_interfaces_http.ApplicationResponse:
    properties:
      artifact_type:
        type: string
      created_at:
        type: string
      description:
        type: string
      gitops_branch_mappings:
        items:
          $ref: '#/definitions/gos_internal_domain_application.GitOpsBranchMapping'
        type: array
      id:
        type: string
      key:
        type: string
      language:
        type: string
      name:
        type: string
      owner:
        type: string
      owner_user_id:
        type: string
      project_id:
        type: string
      project_key:
        type: string
      project_name:
        type: string
      release_branches:
        items:
          $ref: '#/definitions/gos_internal_domain_application.ReleaseBranchOption'
        type: array
      repo_url:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  internal_interfaces_http.CreateApplicationRequest:
    properties:
      artifact_type:
        type: string
      description:
        type: string
      gitops_branch_mappings:
        items:
          $ref: '#/definitions/gos_internal_domain_application.GitOpsBranchMapping'
        type: array
      key:
        type: string
      language:
        type: string
      name:
        type: string
      owner:
        type: string
      owner_user_id:
        type: string
      project_id:
        type: string
      release_branches:
        items:
          $ref: '#/definitions/gos_internal_domain_application.ReleaseBranchOption'
        type: array
      repo_url:
        type: string
      status:
        type: string
    type: object
...
  internal_interfaces_http.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  internal_interfaces_http.GenericResponse:
    properties:
      data:
        additionalProperties: true
        type: object
    type: object
...
  internal_interfaces_http.UpdateApplicationRequest:
    properties:
      artifact_type:
        type: string
      description:
        type: string
      gitops_branch_mappings:
        items:
          $ref: '#/definitions/gos_internal_domain_application.GitOpsBranchMapping'
        type: array
      key:
        type: string
      language:
        type: string
      name:
        type: string
      owner:
        type: string
      owner_user_id:
        type: string
      project_id:
        type: string
      release_branches:
        items:
          $ref: '#/definitions/gos_internal_domain_application.ReleaseBranchOption'
        type: array
      repo_url:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
  description: Internal deployment platform API.
  title: GOS Release API
  version: "1.0"
paths:
...
  /applications:
    get:
      parameters:
      - description: Application keyword (name or key)
        in: query
        name: keyword
        type: string
      - description: Application key
        in: query
        name: key
        type: string
      - description: Application name
        in: query
        name: name
        type: string
      - description: Project ID
        in: query
        name: project_id
        type: string
      - description: Application status
        in: query
        name: status
        type: string
      - description: Page number, starts from 1
        in: query
        name: page
        type: integer
      - description: Page size, max 100
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ApplicationListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: List applications
      tags:
      - applications
    post:
      consumes:
      - application/json
      parameters:
      - description: Create application request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http.CreateApplicationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_interfaces_http.ApplicationDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: Create application
      tags:
      - applications
  /applications/{id}:
    delete:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: Delete application
      tags:
      - applications
    get:
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ApplicationDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: Get application by ID
      tags:
      - applications
    put:
      consumes:
      - application/json
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Update application request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http.UpdateApplicationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.ApplicationDataResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: Update application
      tags:
      - applications
...
  /applications/options:
    get:
      description: 查询Options列表，并按统一响应结构返回处理结果。
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interfaces_http.GenericResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_interfaces_http.ErrorResponse'
      summary: 查询Options列表
      tags:
      - applications
...
schemes:
- http
- https
swagger: "2.0"